Refactor for Error Mesaage class use
